requirements.txt
absl-py==0.12.0
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.3
bleach==3.3.0
blinker==1.4
bs4==0.0.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==8.0.0
cloudpickle==1.6.0
colorama==0.4.4
confuse==1.3.0
cycler==0.10.0
Cython==0.29.23
decorator==4.4.2
defusedxml==0.7.1
-e git+https://github.com/augmentedstartups/detectron2@dcf717d7b256c08a4b2c1fb4d98a84850da6e0f0#egg=detectron2
easyocr==1.3.1
entrypoints==0.3
future==0.18.2
fvcore @ git+https://github.com/facebookresearch/fvcore.git@a68f14296425729d002f212a05a186814eac97ca
gitdb==4.0.7
GitPython==3.1.16
google-auth==1.29.0
google-auth-oauthlib==0.4.4
grpcio==1.37.0
idna==2.10
ImageHash==4.1.0
imageio==2.9.0
importlib-metadata==4.0.1
iopath==0.1.8
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.2
missingno==0.4.2
mistune==0.8.4
mkl-fft==1.3.0
mkl-random @ file:///C:/ci/mkl_random_1618854593605/work
mkl-service==2.3.0
mock==4.0.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.3.0
numpy==1.20.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
packaging==20.9
pandas==1.1.5
pandas-profiling==2.9.0
pandocfilters==1.4.3
parso==0.8.2
phik==0.10.0
pickleshare==0.7.5
Pillow @ file:///C:/ci/pillow_1617386319301/work
portalocker==2.3.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.15.8
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydeck==0.6.2
pydot==1.4.2
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scipy==1.6.3
seaborn==0.11.0
selenium==3.141.0
Send2Trash==1.5.0
six @ file:///C:/ci/six_1605205426665/work
smmap==4.0.0
soupsieve==2.2.1
streamlit==0.81.1
tabulate==0.8.9
tangled-up-in-unicode==0.0.6
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
tifffile==2021.4.8
toml==0.10.2
toolz==0.11.1
torch==1.6.0
torchvision==0.7.0
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
validators==0.18.2
visions==0.5.0
watchdog==2.1.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
yacs==0.1.8
zipp==3.4.1
